move settings and compare
