+++ som_gui/core/filter_window.py
@@ -4,10 +4,13 @@ import logging
 from typing import TYPE_CHECKING, Type
 import SOMcreator
 from PySide6.QtCore import Qt, QItemSelection, QModelIndex
+from PySide6.QtWidgets import QFormLayout, QLabel, QCheckBox
+from PySide6.QtGui import QIcon
 
 if TYPE_CHECKING:
     from som_gui import tool
-
+    from som_gui.module.attribute import ui as attribute_ui
+    from som_gui.module.filter_window import ui
 
 def open_window(filter_window: Type[tool.FilterWindow], project: Type[tool.Project], util: Type[tool.Util],
                 search: Type[tool.Search]):
@@ -97,3 +100,147 @@ def tree_mouse_release_event(index: QModelIndex, filter_window: Type[tool.Filter
     if index is None:
         return
     filter_window.tree_release_click_drag()
+
+
+def add_compare_widget(filter_compare: Type[tool.FilterCompare],
+                       attribute_compare: Type[tool.AttributeCompare],
+                       compare_window: Type[tool.CompareWindow]):
+    compare_window.add_tab("Objekt Filter", filter_compare.get_widget,
+                           lambda p0, p1: init_compare_object_filter(p0, p1, filter_compare, attribute_compare),
+                           filter_compare,
+                           lambda file: export_filter_differences(file, filter_compare, attribute_compare))
+
+
+def init_compare_object_filter(project0: SOMcreator.Project, project1: SOMcreator.Project,
+                               filter_compare: Type[tool.FilterCompare],
+                               attribute_compare: Type[tool.AttributeCompare]):
+    attribute_compare.set_projects(project0, project1)
+    filter_compare.set_projects(project0, project1)
+    attribute_compare.create_object_lists()
+
+    widget = filter_compare.get_widget()
+    object_tree_widget = attribute_compare.get_object_tree(widget)
+    pset_tree = attribute_compare.get_pset_tree(widget)
+    value_table = attribute_compare.get_value_table(widget)
+    filter_compare.make_header_wordwrap(object_tree_widget)
+    filter_compare.make_header_wordwrap(pset_tree)
+
+    attribute_compare.fill_object_tree(object_tree_widget, add_missing=False)
+    header_labels = [attribute_compare.get_header_name_from_project(project0),
+                     attribute_compare.get_header_name_from_project(project1)]
+    attribute_compare.set_header_labels([object_tree_widget, pset_tree], [value_table], header_labels)
+    filter_compare.create_tree_selection_trigger(widget)
+    filter_compare.find_matching_phases(project0, project1)
+    filter_compare.find_matching_usecases(project0, project1)
+
+    filter_compare.append_collumns(object_tree_widget, pset_tree)
+    for child_index in range(object_tree_widget.invisibleRootItem().childCount()):
+        child = object_tree_widget.invisibleRootItem().child(child_index)
+        filter_compare.fill_tree_with_checkstates(child)
+        filter_compare.style_object_tree(child)
+    for col in range(2, object_tree_widget.columnCount()):
+        object_tree_widget.setColumnWidth(col, 58)
+
+    widget.widget.table_widget_values.hide()
+
+
+def filter_tab_object_tree_selection_changed(widget: attribute_ui.AttributeWidget,
+                                             attribute_compare: Type[tool.AttributeCompare],
+                                             filter_compare: Type[tool.FilterCompare]):
+    obj = attribute_compare.get_selected_entity(attribute_compare.get_object_tree(widget))
+    tree_widget = attribute_compare.get_pset_tree(widget)
+    pset_list = attribute_compare.get_pset_list(obj)
+
+    attribute_compare.fill_pset_tree(tree_widget, pset_list, add_missing=False)
+    attribute_compare.add_attributes_to_pset_tree(tree_widget, False)
+    for child_index in range(tree_widget.invisibleRootItem().childCount()):
+        child = tree_widget.invisibleRootItem().child(child_index)
+        filter_compare.fill_tree_with_checkstates(child)
+
+    for col in range(2, tree_widget.columnCount()):
+        tree_widget.setColumnWidth(col, 58)
+
+
+def export_filter_differences(file, filter_compare: Type[tool.FilterCompare],
+                              attribute_compare: Type[tool.AttributeCompare]):
+    file.write(f"\n{'OBJECT FILTER':46s}\n\n")
+    filter_compare.export_object_filter_differences(file, attribute_compare)
+    file.write("\n")
+
+
+# Settings WIdget
+def settings_widget_created(widget: ui.SettingsWidget, filter_window: Type[tool.FilterWindow],
+                            project: Type[tool.Project]):
+    filter_window.set_settings_widget(widget)
+
+    proj = project.get()
+    phase_layout = QFormLayout()
+    widget.ui.widget_phase.setLayout(phase_layout)
+    usecase_layout = QFormLayout()
+    widget.ui.widget_usecase.setLayout(usecase_layout)
+
+    for phase in proj.get_phases():
+        cb = QCheckBox()
+        phase_index = proj.get_phase_index(phase)
+        cb.setChecked(bool(phase_index in proj.active_phases))
+        phase_layout.addRow(QLabel(phase.name), cb)
+
+    for usecase in proj.get_usecases():
+        cb = QCheckBox()
+        usecase_index = proj.get_use_case_index(usecase)
+        cb.setChecked(bool(usecase_index in proj.active_usecases))
+        usecase_layout.addRow(QLabel(usecase.name), cb)
+
+
+def settings_accepted(filter_window: Type[tool.FilterWindow], project: Type[tool.Project], popups: Type[tool.Popups]):
+    proj = project.get()
+    widget = filter_window.get_settings_widget()
+    phase_layout: QFormLayout = widget.ui.widget_phase.layout()
+    usecase_layout: QFormLayout = widget.ui.widget_usecase.layout()
+
+    active_phases = list()
+    for row in range(phase_layout.rowCount()):
+        cb: QCheckBox = phase_layout.itemAt(row, QFormLayout.ItemRole.FieldRole).widget()
+        if cb.isChecked():
+            active_phases.append(row)
+
+    active_usecases = list()
+    for row in range(usecase_layout.rowCount()):
+        cb: QCheckBox = usecase_layout.itemAt(row, QFormLayout.ItemRole.FieldRole).widget()
+        if cb.isChecked():
+            active_usecases.append(row)
+    proj.active_phases = active_phases
+    proj.active_usecases = active_usecases
+    logging.info(f"Set Active Usecases {[proj.get_usecase_by_index(i).name for i in proj.active_usecases]}")
+    logging.info(f"Set Active Phases {[proj.get_phase_by_index(i).name for i in proj.active_phases]}")
+
+
+def settings_combobox_changed(filter_window: Type[tool.FilterWindow], project: Type[tool.Project],
+                              util: Type[tool.Util]):
+    widget = filter_window.get_settings_widget()
+    combobox_usecase = widget.ui.cb_usecase
+    combobox_phase = widget.ui.cb_phase
+    usecase_name = combobox_usecase.currentText()
+    phase_name = combobox_phase.currentText()
+    if not all([usecase_name, phase_name]):
+        return
+    proj = project.get()
+    usecase = proj.get_use_case_by_name(usecase_name)
+    phase = proj.get_phase_by_name(phase_name)
+
+    # add warning icons to combobox
+    for uc_name, index in util.get_text_from_combobox(combobox_usecase).items():
+        uc = proj.get_use_case_by_name(uc_name)
+        if not proj.get_filter_state(phase, uc):
+            warn_icon = widget.style().standardIcon(widget.style().StandardPixmap.SP_MessageBoxWarning)
+            combobox_usecase.setItemIcon(index.row(), warn_icon)
+        else:
+            combobox_usecase.setItemIcon(index.row(), QIcon())
+
+    for ph_name, index in util.get_text_from_combobox(combobox_phase).items():
+        ph = proj.get_phase_by_name(ph_name)
+        if not proj.get_filter_state(ph, usecase):
+            warn_icon = widget.style().standardIcon(widget.style().StandardPixmap.SP_MessageBoxWarning)
+            combobox_phase.setItemIcon(index.row(), warn_icon)
+        else:
+            combobox_phase.setItemIcon(index.row(), QIcon())
