+++ som_gui/module/filter_window/constants.py
