+++ som_gui/module/filter_window/qt/Settings.ui
