+++ som_gui/module/filter_window/qt/settings.py
