+++ som_gui/module/filter_window/trigger.py
@@ -4,12 +4,14 @@ from som_gui import tool
 from som_gui.core import filter_window as core
 from PySide6.QtCore import Qt
 from typing import TYPE_CHECKING
-
+from . import ui, constants
 
 def connect():
     tool.MainWindow.add_action("Bearbeiten",
                                lambda: core.open_window(tool.FilterWindow, tool.Project, tool.Util, tool.Search))
-
+    core.add_compare_widget(tool.FilterCompare, tool.AttributeCompare, tool.CompareWindow)
+    tool.Settings.add_page_to_toolbox(ui.SettingsWidget, constants.SETTINGS_TAB_NAME, constants.SETTINGS_PAGE_NAME,
+                                      lambda: core.settings_accepted(tool.FilterWindow, tool.Project, tool.Popups))
 
 
 def pt_horizontal_context_requested(pos):
@@ -46,3 +48,17 @@ def object_tree_clicked(selected, deselected):
 
 def update_pset_tree():
     core.update_pset_tree(tool.FilterWindow)
+
+
+# Settings
+
+def settings_widget_created(widget: ui.SettingsWidget):
+    core.settings_widget_created(widget, tool.FilterWindow, tool.Project)
+
+
+def settings_combobox_changed():
+    core.settings_combobox_changed(tool.FilterWindow, tool.Project, tool.Util)
+
+
+def filter_tab_object_tree_selection_changed(widget):
+    core.filter_tab_object_tree_selection_changed(widget, tool.AttributeCompare, tool.FilterCompare)
