--- som_gui/module/object_filter/qt/widget.py
@@ -1,73 +1,0 @@
-# -*- coding: utf-8 -*-
-
-################################################################################
-## Form generated from reading UI file 'UseCaseWindow.ui'
-##
-## Created by: Qt User Interface Compiler version 6.6.1
-##
-## WARNING! All changes made in this file will be lost when recompiling UI file!
-################################################################################
-
-from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
-    QMetaObject, QObject, QPoint, QRect,
-    QSize, QTime, QUrl, Qt)
-from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
-    QFont, QFontDatabase, QGradient, QIcon,
-    QImage, QKeySequence, QLinearGradient, QPainter,
-    QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QAbstractButton, QAbstractItemView, QApplication, QDialogButtonBox,
-    QHeaderView, QLabel, QSizePolicy, QSplitter,
-    QVBoxLayout, QWidget)
-
-from som_gui.module.object_filter.ui import ObjectTreeView
-
-class Ui_Form(object):
-    def setupUi(self, Form):
-        if not Form.objectName():
-            Form.setObjectName(u"Form")
-        Form.resize(1341, 581)
-        self.verticalLayout = QVBoxLayout(Form)
-        self.verticalLayout.setObjectName(u"verticalLayout")
-        self.splitter = QSplitter(Form)
-        self.splitter.setObjectName(u"splitter")
-        self.splitter.setOrientation(Qt.Horizontal)
-        self.object_tree = ObjectTreeView(self.splitter)
-        self.object_tree.setObjectName(u"object_tree")
-        self.object_tree.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        self.splitter.addWidget(self.object_tree)
-        self.verticalLayoutWidget = QWidget(self.splitter)
-        self.verticalLayoutWidget.setObjectName(u"verticalLayoutWidget")
-        self.verticalLayout_2 = QVBoxLayout(self.verticalLayoutWidget)
-        self.verticalLayout_2.setObjectName(u"verticalLayout_2")
-        self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.label_object = QLabel(self.verticalLayoutWidget)
-        self.label_object.setObjectName(u"label_object")
-
-        self.verticalLayout_2.addWidget(self.label_object)
-
-        self.property_set_tree = ObjectTreeView(self.verticalLayoutWidget)
-        self.property_set_tree.setObjectName(u"property_set_tree")
-
-        self.verticalLayout_2.addWidget(self.property_set_tree)
-
-        self.splitter.addWidget(self.verticalLayoutWidget)
-
-        self.verticalLayout.addWidget(self.splitter)
-
-        self.buttonBox = QDialogButtonBox(Form)
-        self.buttonBox.setObjectName(u"buttonBox")
-        self.buttonBox.setStandardButtons(QDialogButtonBox.Cancel|QDialogButtonBox.Ok)
-
-        self.verticalLayout.addWidget(self.buttonBox)
-
-
-        self.retranslateUi(Form)
-
-        QMetaObject.connectSlotsByName(Form)
-    # setupUi
-
-    def retranslateUi(self, Form):
-        Form.setWindowTitle(QCoreApplication.translate("Form", u"Anwendungsf\u00e4lle", None))
-        self.label_object.setText(QCoreApplication.translate("Form", u"TextLabel", None))
-    # retranslateUi
-
