--- som_gui/module/project_filter/qt/widget.py
@@ -1,59 +1,0 @@
-# -*- coding: utf-8 -*-
-
-################################################################################
-## Form generated from reading UI file 'project_filter.ui'
-##
-## Created by: Qt User Interface Compiler version 6.7.0
-##
-## WARNING! All changes made in this file will be lost when recompiling UI file!
-################################################################################
-
-from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
-                            QMetaObject, QObject, QPoint, QRect,
-                            QSize, QTime, QUrl, Qt)
-from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
-                           QFont, QFontDatabase, QGradient, QIcon,
-                           QImage, QKeySequence, QLinearGradient, QPainter,
-                           QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QAbstractButton, QAbstractItemView, QApplication, QDialog,
-                               QDialogButtonBox, QGridLayout, QHeaderView, QSizePolicy,
-                               QTableWidgetItem, QWidget)
-
-from som_gui.module.project_filter.ui import ProjectFilterTable
-
-class Ui_Dialog(object):
-    def setupUi(self, Dialog):
-        if not Dialog.objectName():
-            Dialog.setObjectName(u"Dialog")
-        Dialog.resize(1047, 698)
-        self.gridLayout = QGridLayout(Dialog)
-        self.gridLayout.setObjectName(u"gridLayout")
-        self.tableWidget = ProjectFilterTable(Dialog)
-        self.tableWidget.setObjectName(u"tableWidget")
-        self.tableWidget.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        self.tableWidget.setDragEnabled(False)
-        self.tableWidget.setDragDropOverwriteMode(False)
-        self.tableWidget.setDragDropMode(QAbstractItemView.NoDragDrop)
-        self.tableWidget.horizontalHeader().setStretchLastSection(True)
-
-        self.gridLayout.addWidget(self.tableWidget, 0, 0, 1, 1)
-
-        self.buttonBox = QDialogButtonBox(Dialog)
-        self.buttonBox.setObjectName(u"buttonBox")
-        self.buttonBox.setOrientation(Qt.Horizontal)
-        self.buttonBox.setStandardButtons(QDialogButtonBox.Ok)
-        self.buttonBox.setCenterButtons(False)
-
-        self.gridLayout.addWidget(self.buttonBox, 1, 0, 1, 1)
-
-        self.retranslateUi(Dialog)
-        self.buttonBox.accepted.connect(Dialog.accept)
-        self.buttonBox.rejected.connect(Dialog.reject)
-
-        QMetaObject.connectSlotsByName(Dialog)
-    # setupUi
-
-    def retranslateUi(self, Dialog):
-        Dialog.setWindowTitle(QCoreApplication.translate("Dialog", u"Dialog", None))
-    # retranslateUi
-
